minor tweaks and bug fixes

+++ switch_model/upgrade/upgrade_2_0_0b4.py
@@ -25,19 +25,19 @@ def upgrade_input_dir(inputs_dir):
         if optional_file and not os.path.isfile(old_path):
             return
         shutil.move(old_path, new_path)
-    
+
     def rename_column(file_name, old_col_name, new_col_name, optional_file=True):
         path = os.path.join(inputs_dir, file_name)
         if optional_file and not os.path.isfile(path):
             return
         df = pandas.read_csv(path, na_values=['.'], sep='\t')
         df.rename(columns={old_col_name: new_col_name}, inplace=True)
-        df.to_csv(path, sep='\t', na_rep='.', index=False)        
-        
+        df.to_csv(path, sep='\t', na_rep='.', index=False)
+
     old_new_column_names_in_file = {
         'gen_inc_heat_rates.tab': [('project', 'GENERATION_PROJECT')]
     }
-    
+
     for fname, old_new_pairs in old_new_column_names_in_file.iteritems():
         for old, new in old_new_pairs:
             rename_column(fname, old_col_name=old, new_col_name=new)
@@ -47,10 +47,12 @@ def upgrade_input_dir(inputs_dir):
     trans_opt_path = os.path.join(inputs_dir, 'trans_optional_params.tab')
     if os.path.isfile(trans_lines_path) and os.path.isfile(trans_lines_path):
         trans_lines = pandas.read_csv(trans_lines_path, na_values=['.'], sep='\t')
-        trans_opt = pandas.read_csv(trans_opt_path, na_values=['.'], sep='\t')
-        trans_lines = trans_lines.merge(trans_opt, on='TRANSMISSION_LINE', how='left')
+        if os.path.isfile(trans_opt_path):
+            trans_opt = pandas.read_csv(trans_opt_path, na_values=['.'], sep='\t')
+            trans_lines = trans_lines.merge(trans_opt, on='TRANSMISSION_LINE', how='left')
         trans_lines.to_csv(trans_lines_path, sep='\t', na_rep='.', index=False)
-        os.remove(trans_opt_path)
+        if os.path.isfile(trans_opt_path):
+            os.remove(trans_opt_path)
 
     # Write a new version text file.
     switch_model.upgrade._write_input_version(inputs_dir, upgrades_to)
